requirements.txt
Django==5.0.3
django-timezone-field==6.1.0
proglog==0.1.10
pyserial==3.5

